Refactor developer context
