+++ assets/src/pages/home/HomePage.tsx
@@ -1,7 +1,7 @@
-import React, { FC, useContext, useState } from "react";
+import React, { FC, useState } from "react";
 import { Link, useParams, useSearchParams } from "react-router-dom";
 import clsx from "clsx";
-import { DeveloperContext } from "../../contexts/developerContext";
+import { useDeveloperInfo } from "../../contexts/DeveloperInfoContext";
 import { Checkbox, Props as CheckboxProps } from "./Checkbox";
 import { useToggle } from "../room/hooks/useToggle";
 import { useMediaDeviceManager } from "../room/hooks/useMediaDeviceManager";
@@ -16,7 +16,7 @@ export const HomePage: FC = () => {
 
   const match = useParams();
   const [searchParams] = useSearchParams();
-  const { manualMode, simulcast, cameraAutostart } = useContext(DeveloperContext);
+  const { manualMode, simulcast, cameraAutostart } = useDeveloperInfo();
   const { setUsername } = useUser();
 
   const roomId: string = match?.roomId || "";
