+++ assets/src/pages/room/consts.ts
@@ -34,6 +34,13 @@ export const VIDEO_TRACKS_CONFIG = new MediaStreamConfig(VIDEO_TRACK_CONSTRAINTS
 export const AUDIO_TRACKS_CONFIG = new MediaStreamConfig(AUDIO_TRACK_CONSTRAINTS);
 export const SCREEN_SHARING_TRACKS_CONFIG = new DisplayMediaStreamConfig(SCREENSHARING_MEDIA_CONSTRAINTS);
 
+export const LOCAL_PEER_NAME = "Me"
+export const LOCAL_VIDEO_ID = "LOCAL_VIDEO_ID"
+export const LOCAL_SCREEN_SHARING_ID = "LOCAL_SCREEN_SHARING_ID"
+
+export const DEFAULT_AUTOSTART_CAMERA_AND_MICROPHONE_CHECKBOX_VALUE = true;
+export const DEFAULT_MANUAL_MODE_CHECKBOX_VALUE = false;
+
 // todo implement
 const MBps = 1024 * 8;
